+++ ChoGGi's Library/Code/Classes_UI.lua
@@ -199,6 +199,20 @@ function ChoGGi_XMoveControl:OnMouseButtonDoubleClick(...)
 	return XMoveControl.OnMouseButtonDoubleClick(self, ...)
 end
 
+function ChoGGi_XMoveControl:OnMouseButtonDown(pt, button, ...)
+	-- make selected window the foremost window
+  if button == "L" then
+		-- default is 5, make them all 5 and selected to 6
+		local ChoGGi_dlgs_opened = ChoGGi_dlgs_opened
+		for win in pairs(ChoGGi_dlgs_opened) do
+			win:SetZOrder(5)
+		end
+		self.parent_dialog:SetZOrder(6)
+	end
+
+	return XMoveControl.OnMouseButtonDown(self, pt, button, ...)
+end
+
 DefineClass.ChoGGi_XSizeControl = {
 	__parents = {
 		"ChoGGi_XDefaults",