+++ Mods ChoGGi/Catacombes de Mars/metadata.lua
@@ -1,0 +1,26 @@
+return PlaceObj("ModDef", {
+	"title", "Catacombes de Mars",
+	"id", "ChoGGi_CatacombesdeMars",
+	"lua_revision", 249143,
+	"steam_id", "2170232881",
+	"pops_any_uuid", "74704c3f-4cde-4ed7-8b17-7c8bb0753a43",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.png",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+--~ 	"has_options", true,
+	"TagGameplay", true,
+	"TagBuildings", true,
+	"description", [[Tunnels are now interconnected.
+
+Lightly tested, so there could be issues.
+I only bothered testing with rovers, no clue if elec/life support works.
+
+
+Requested by Llamacorn1921.
+]],
+})